center finding bug fix, clutch :) clustering works too. Getting closer!!

+++ src/AnalyzeFrame.java
@@ -393,19 +393,19 @@ public class AnalyzeFrame {
 		zr = Z/Zr;
 
 		if ( xr > 0.008856 )
-			xr =  (float) Math.pow(xr, 1/3.);
+			xr =  (double) Math.pow(xr, 1/3.);
 		else
-			xr = (float) ((7.787 * xr) + 16 / 116.0);
+			xr = (double) ((7.787 * xr) + 16 / 116.0);
 
 		if ( yr > 0.008856 )
-			yr =  (float) Math.pow(yr, 1/3.);
+			yr =  (double) Math.pow(yr, 1/3.);
 		else
-			yr = (float) ((7.787 * yr) + 16 / 116.0);
+			yr = (double) ((7.787 * yr) + 16 / 116.0);
 
 		if ( zr > 0.008856 )
-			zr =  (float) Math.pow(zr, 1/3.);
+			zr =  (double) Math.pow(zr, 1/3.);
 		else
-			zr = (float) ((7.787 * zr) + 16 / 116.0);
+			zr = (double) ((7.787 * zr) + 16 / 116.0);
 
 
 		double[] lab = new double[3];
